pwm update

+++ Edison/pwm/pwm.py
@@ -26,10 +26,10 @@ while True:
     results_raw = requests.get("http://fryer.ee.ucla.edu/rest/api/pwm/get/")
     results_json = results_raw.text
     results = yaml.safe_load(results_json)
-    pwm1_value = double(results['data']['pwm1'])
-    pwm2_value = double(results['data']['pwm2'])
-    pwm3_value = double(results['data']['pwm3'])
-    pwm4_value = double(results['data']['pwm4'])
+    pwm1_value = float(results['data']['pwm1'])
+    pwm2_value = float(results['data']['pwm2'])
+    pwm3_value = float(results['data']['pwm3'])
+    pwm4_value = float(results['data']['pwm4'])
 
     while value<0.1:
           pwm1.write(pwm1_value)
